# Conflicts:
#	Dockerfile
#	src/Database/Database.php

+++ src/Database/Adapter/MariaDB.php
@@ -146,8 +146,8 @@ class MariaDB extends SQL
             $indexStrings[$key] = "{$indexType} `{$indexId}` ({$indexAttributes}),";
         }
 
-        $sql = "
-			CREATE TABLE IF NOT EXISTS {$this->getSQLTable($id)} (
+        $collection = "
+			CREATE TABLE {$this->getSQLTable($id)} (
 				_id INT(11) UNSIGNED NOT NULL AUTO_INCREMENT,
 				_uid VARCHAR(255) NOT NULL,
 				_createdAt DATETIME(3) DEFAULT NULL,
@@ -159,7 +159,7 @@ class MariaDB extends SQL
 		";
 
         if ($this->sharedTables) {
-            $sql .= "
+            $collection .= "
             	_tenant INT(11) UNSIGNED DEFAULT NULL,
 				UNIQUE KEY _uid (_tenant, _uid),
 				KEY _created_at (_tenant, _createdAt),
@@ -167,56 +167,59 @@ class MariaDB extends SQL
 				KEY _tenant_id (_tenant, _id)
 			";
         } else {
-            $sql .= "
+            $collection .= "
 				UNIQUE KEY _uid (_uid),
 				KEY _created_at (_createdAt),
 				KEY _updated_at (_updatedAt)
 			";
         }
 
-        $sql .= ")";
+        $collection .= ")";
+        $collection = $this->trigger(Database::EVENT_COLLECTION_CREATE, $collection);
 
-        $sql = $this->trigger(Database::EVENT_COLLECTION_CREATE, $sql);
+        $permissions = "
+            CREATE TABLE {$this->getSQLTable($id . '_perms')} (
+                _id int(11) UNSIGNED NOT NULL AUTO_INCREMENT,
+                _type VARCHAR(12) NOT NULL,
+                _permission VARCHAR(255) NOT NULL,
+                _document VARCHAR(255) NOT NULL,
+                PRIMARY KEY (_id),
+        ";
+
+        if ($this->sharedTables) {
+            $permissions .= "
+                _tenant INT(11) UNSIGNED DEFAULT NULL,
+                UNIQUE INDEX _index1 (_document, _tenant, _type, _permission),
+                INDEX _permission (_tenant, _permission, _type)
+            ";
+        } else {
+            $permissions .= "
+                UNIQUE INDEX _index1 (_document, _type, _permission),
+                INDEX _permission (_permission, _type)
+            ";
+        }
+
+        $permissions .= ")";
+        $permissions = $this->trigger(Database::EVENT_COLLECTION_CREATE, $permissions);
 
         try {
             $this->getPDO()
-                ->prepare($sql)
+                ->prepare($collection)
                 ->execute();
 
-            $sql = "
-				CREATE TABLE IF NOT EXISTS {$this->getSQLTable($id . '_perms')} (
-					_id int(11) UNSIGNED NOT NULL AUTO_INCREMENT,
-					_type VARCHAR(12) NOT NULL,
-					_permission VARCHAR(255) NOT NULL,
-					_document VARCHAR(255) NOT NULL,
-					PRIMARY KEY (_id),
-			";
+            $this->getPDO()
+                ->prepare($permissions)
+                ->execute();
+        } catch (PDOException $e) {
+            $e = $this->processException($e);
 
-            if ($this->sharedTables) {
-                $sql .= "
-                	_tenant INT(11) UNSIGNED DEFAULT NULL,
-					UNIQUE INDEX _index1 (_document, _tenant, _type, _permission),
-					INDEX _permission (_tenant, _permission, _type)
-				";
-            } else {
-                $sql .= "
-					UNIQUE INDEX _index1 (_document, _type, _permission),
-					INDEX _permission (_permission, _type)
-				";
+            if (!($e instanceof DuplicateException)) {
+                $this->getPDO()
+                    ->prepare("DROP TABLE IF EXISTS {$this->getSQLTable($id)}, {$this->getSQLTable($id . '_perms')};")
+                    ->execute();
             }
 
-            $sql .= ")";
-
-            $sql = $this->trigger(Database::EVENT_COLLECTION_CREATE, $sql);
-
-            $this->getPDO()
-                ->prepare($sql)
-                ->execute();
-        } catch (\Exception $th) {
-            $this->getPDO()
-                ->prepare("DROP TABLE IF EXISTS {$this->getSQLTable($id)}, {$this->getSQLTable($id . '_perms')};")
-                ->execute();
-            throw $th;
+            throw $e;
         }
 
         return true;
@@ -1076,7 +1079,7 @@ class MariaDB extends SQL
 			";
 
             if ($this->sharedTables) {
-                $sql .= ' AND _tenant = :_tenant';
+                $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
             }
 
             $sql = $this->trigger(Database::EVENT_PERMISSIONS_READ, $sql);
@@ -1153,7 +1156,7 @@ class MariaDB extends SQL
                 ";
 
                 if ($this->sharedTables) {
-                    $sql .= ' AND _tenant = :_tenant';
+                    $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
                 }
 
                 $removeQuery = $sql . $removeQuery;
@@ -1240,7 +1243,7 @@ class MariaDB extends SQL
 			";
 
             if ($this->sharedTables) {
-                $sql .= ' AND _tenant = :_tenant';
+                $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
             }
 
             $sql = $this->trigger(Database::EVENT_DOCUMENT_UPDATE, $sql);
@@ -1369,7 +1372,7 @@ class MariaDB extends SQL
                     ";
 
                     if ($this->sharedTables) {
-                        $sql .= ' AND _tenant = :_tenant';
+                        $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
                     }
 
                     $sql = $this->trigger(Database::EVENT_PERMISSIONS_READ, $sql);
@@ -1412,7 +1415,7 @@ class MariaDB extends SQL
 
                             $tenantQuery = '';
                             if ($this->sharedTables) {
-                                $tenantQuery = ' AND _tenant = :_tenant';
+                                $tenantQuery = ' AND (_tenant = :_tenant OR _tenant IS NULL)';
                             }
 
                             $removeQuery .= "(
@@ -1570,12 +1573,13 @@ class MariaDB extends SQL
      * @param string $id
      * @param string $attribute
      * @param int|float $value
+     * @param string $updatedAt
      * @param int|float|null $min
      * @param int|float|null $max
      * @return bool
-     * @throws Exception
+     * @throws DatabaseException
      */
-    public function increaseDocumentAttribute(string $collection, string $id, string $attribute, int|float $value, int|float|null $min = null, int|float|null $max = null): bool
+    public function increaseDocumentAttribute(string $collection, string $id, string $attribute, int|float $value, string $updatedAt, int|float|null $min = null, int|float|null $max = null): bool
     {
         $name = $this->filter($collection);
         $attribute = $this->filter($attribute);
@@ -1585,12 +1589,14 @@ class MariaDB extends SQL
 
         $sql = "
 			UPDATE {$this->getSQLTable($name)} 
-			SET `{$attribute}` = `{$attribute}` + :val 
+			SET 
+			    `{$attribute}` = `{$attribute}` + :val,
+			    `_updatedAt` = :updatedAt
 			WHERE _uid = :_uid
 		";
 
         if ($this->sharedTables) {
-            $sql .= ' AND _tenant = :_tenant';
+            $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
         }
 
         $sql .= $sqlMax . $sqlMin;
@@ -1600,6 +1606,7 @@ class MariaDB extends SQL
         $stmt = $this->getPDO()->prepare($sql);
         $stmt->bindValue(':_uid', $id);
         $stmt->bindValue(':val', $value);
+        $stmt->bindValue(':updatedAt', $updatedAt);
 
         if ($this->sharedTables) {
             $stmt->bindValue(':_tenant', $this->tenant);
@@ -1631,7 +1638,7 @@ class MariaDB extends SQL
 		    ";
 
             if ($this->sharedTables) {
-                $sql .= ' AND _tenant = :_tenant';
+                $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
             }
 
             $sql = $this->trigger(Database::EVENT_DOCUMENT_DELETE, $sql);
@@ -1650,7 +1657,7 @@ class MariaDB extends SQL
 		    ";
 
             if ($this->sharedTables) {
-                $sql .= ' AND _tenant = :_tenant';
+                $sql .= ' AND (_tenant = :_tenant OR _tenant IS NULL)';
             }
 
             $sql = $this->trigger(Database::EVENT_PERMISSIONS_DELETE, $sql);
@@ -1700,6 +1707,7 @@ class MariaDB extends SQL
      * @return array<Document>
      * @throws DatabaseException
      * @throws TimeoutException
+     * @throws Exception
      */
     public function find(string $collection, array $queries = [], ?int $limit = 25, ?int $offset = null, array $orderAttributes = [], array $orderTypes = [], array $cursor = [], string $cursorDirection = Database::CURSOR_AFTER): array
     {
@@ -1793,7 +1801,7 @@ class MariaDB extends SQL
         }
 
         if ($this->sharedTables) {
-            $where[] = "table_main._tenant = :_tenant";
+            $where[] = "table_main._tenant IN (:_tenant, NULL)";
         }
 
         $sqlWhere = !empty($where) ? 'WHERE ' . implode(' AND ', $where) : '';
@@ -1850,7 +1858,7 @@ class MariaDB extends SQL
         try {
             $stmt->execute();
         } catch (PDOException $e) {
-            $this->processException($e);
+            throw $this->processException($e);
         }
 
         $results = $stmt->fetchAll();
@@ -1919,7 +1927,7 @@ class MariaDB extends SQL
         }
 
         if ($this->sharedTables) {
-            $where[] = "table_main._tenant = :_tenant";
+            $where[] = "table_main._tenant IN (:_tenant, NULL)";
         }
 
         $sqlWhere = !empty($where)
@@ -1989,7 +1997,7 @@ class MariaDB extends SQL
         }
 
         if ($this->sharedTables) {
-            $where[] = "table_main._tenant = :_tenant";
+            $where[] = "table_main._tenant IN (:_tenant, NULL)";
         }
 
         $sqlWhere = !empty($where)
@@ -2269,22 +2277,22 @@ class MariaDB extends SQL
         });
     }
 
-    /**
-     * @param PDOException $e
-     * @throws TimeoutException
-     */
-    protected function processException(PDOException $e): void
+    protected function processException(PDOException $e): \Exception
     {
-        // Regular PDO
+        // Timeout
         if ($e->getCode() === '70100' && isset($e->errorInfo[1]) && $e->errorInfo[1] === 1969) {
-            throw new TimeoutException($e->getMessage(), $e->getCode(), $e);
+            return new TimeoutException($e->getMessage(), $e->getCode(), $e);
+        } elseif ($e->getCode() === 1969 && isset($e->errorInfo[0]) && $e->errorInfo[0] === '70100') {
+            return new TimeoutException($e->getMessage(), $e->getCode(), $e);
         }
 
-        // PDOProxy switches errorInfo PDOProxy.php line 64
-        if ($e->getCode() === 1969 && isset($e->errorInfo[0]) && $e->errorInfo[0] === '70100') {
-            throw new TimeoutException($e->getMessage(), $e->getCode(), $e);
+        // Duplicate table
+        if ($e->getCode() === '42S01' && isset($e->errorInfo[1]) && $e->errorInfo[1] === 1050) {
+            return new DuplicateException($e->getMessage(), $e->getCode(), $e);
+        } elseif ($e->getCode() === 1050 && isset($e->errorInfo[0]) && $e->errorInfo[0] === '42S01') {
+            return new DuplicateException($e->getMessage(), $e->getCode(), $e);
         }
 
-        throw $e;
+        return $e;
     }
 }
